VERIFICATION: Fixing linter warnings about functions in asserts.
The synopsys linter thinks some functions have side effects and could
affect the outcome of simulations if the asserts were disabled.

The functions in question shouldn't have any side effects and I've run a
number of tests and been unable to write even the simplest function that
it accepts.

As a workaround I've removed function calls from asserts. Which is not
ideal but works.