ISO14443A HDL: Added the Tx module.
It takes the input data and number of bits to send,
and transmits manchester encoded data AND'ed with the subcarrier.
It sends the SOC + data + parity + data + parity + data + ...
It requests new data every time it's running out with the req signal.
It only sends when the fdt trigger asserts.