VERIFICATION: Use blue to indicate assertions that have not been hit.
I've now run into a couple of cases where my generic sinks / interfaces
have assertions that aren't used in particular TBs. Showing them in red,
the same as other errors is not correct since this isn't an actual
error. Blue makes it stand out as something to check, but is easier to
ignore when you know it's fine.