VERIFICATION: Reworked frame_generator_pkg::convert_message_to_bit_queue
Previously you could only specify the number of bits in the last byte.
Which is only useful for Rx comms and not Tx.

Now we have separate functions for Rx and Tx that let you set the number
of bits in the last / first byte, depending on which you use.