VERIFICATION: Increase the time before the REQ comes in after sampling data, in the tx_interface_sink.
I had an error when simulating the new framing module. Since the data
first goes through the serialiser then to the frame encoder before going
to the tx_sink.

If the tx_sink asserts req too early, the data_valid line from the
serialiser is still asserted, so the frame encoder thinks there's
another byte to send.

Asserting req later is a solution, but only if in the final design
nothing asserts req too early. If that's the case we'll see errors in
higher level testbenches.