ISO14443A HDL: Updating initialisation to use both the bit and the byte Rx streams.
The byte stream is used for normal Rx (REQA/WUPA/...).
The bit stream is used for UID matching during AC/SELECT messages.

Previously if we got an AC message with a partial byte, we had to
compare anywhere between 1 and 7 bits of the received data with the
expected UID. This is not a particularly efficient approach.

Now we can wait until we've got the first two bytes of an AC/SELECT message and then start comparing bit by bit with the bit stream.

Additionally to transmit an AC reply we had to take take the entire
actual current uid buffer and then shift out bytes and then bits before
being ready to send. Now we can do this during Rx, so the Tx code is
more efficient too.