VERIFICATION: Starting to rework the verification architecture.
I'm working towards making the verification code more reusable using UVM
concepts.

This is not UVM yet, but it should make converting to UVM later a fair
bit easier.

Basically we now have drivers which take a reference to a send queue of
transactions, and a virtual interface to drive. They wait for a
transaction to appear on the queue and then they set it going.

We also have monitors which monitor a virtual interface and decode
transactions, pushing them to a receive queue as they come in.

So far I'm only supporting the Rx path, but the Tx path will come
shortly.