VERIFICATION: Split the pcd_pause_n_and_clock_source module.
Since we now have a PCDPauseNDriver class which is designed to drive the
pcd_pause_n signal with PCDBitSequences. We no longer need the tasks /
functions in pcd_pause_n_and_clock_source.

We do however need a generic clock source for use with all modules. So
I've added that as it's own module. It just generates a clock of the
desired frequency.

We also need the pause detect part of the pcd_pause_n_and_clock_source,
which i've moved to the pause_detect module. It simply takes the
pcd_pause_n signal as an input and produces a delayed pause_n signal.

Then there is the clock recovery part, which simulates the clock
recovery block of the analogue design. It is simply it's own clock, but
it stops during pause frames to mimic the real design.

Finally I encapsulated all of this, plus the driver and a synchroniser
into the analogue_sim module. This module has three outputs, the picc's
clock, the unsynchronised pause_n signal and a synchronised pause_n
signal. While the pause_n signal is synchronised in the digital part of
this IP core, it's something that most testsbenches that use this will
want. It also contains a couple of functions for interacting with the
driver, but the driver can be directly accessed via hierarchical paths.