VERIFICATION: Added transaction_generators and sequences.
The transaction generators, generate transactions for the common Rx and
Tx messages. Additionally it automatically converts them into the
correct transaction type for the interface, and auto appends CRCs and
adds parity bits as required.

The sequences deal with PCD -> PICC and PICC -> PCD comms. It has
functions to send messages to the PICC and then verify the response.

The SpecificTargetSequence class addss functions to bring a particular
PICC to specific states, using it's UID.

The CommsTests class is the test code ported from the old init_comm_pkg,
and the iso14443_4a_tb. The idea behind merging them is that the
testbenches can pick and choose which tests to run, depending on what is
needed for that testbench.

I've also added several helper classes:

UID: Stores a PICC's UID with helpful get and set methods.

Wrapper: Wraps a type in a class. Since classes are passed and stored
by reference. I use this to store a reference to send and receive
queues.

Target: Stores details about a Target, including UID, block numbers,
CID, ...

StdBlockAddress: When addressing a particular PICC in the protocol
state, you send messages with a header that can optionally including a
CID and a NAD field. The CID and the NAD are the address of the message.
This class stores that data.