VERIFICATION: Added top level testbench.
This tests the entire ISO/IEC 14443A digital IP core.

It's basically a combination of the iso14443_3a_tb and iso14443_4a_tb
tests. It runs all the initialisation and the part4 tests.

It's currently very slow, but that's not that suprising. I intend to
look into optimised builds at some point.