ISO14443A HDL: Updated the top level module.
synchronise the async power input.
renamed some inputs to _async to make it clearer which are the async and
which are the synchronous signals.
Export the part4 <-> app rx and tx interfaces.
Added extensive comments on the FDT timer.
Accounted for delays in the part2 block when passing FDT_TIMING_ADJUST
to the part3 block.
Error if FDT_TIMING_ADJUST is not assigned.