VERIFICATION: Reworked the TransactionGenerators.
Previously there were a heirarchy of TransactionGenerator classes. You
instantiated the correct one and when you called generate_x, it would
generate you a transaction of the correct type (bits, bytes, ...).

The new scheme has a single TransactionGenerator class that crceates all
the ISO/IEC 14443A messages as RxByteTransaction or TxByteTransaction
(depending on which you need). Then there is a separate set of
TransactionConverter classes that convert the byte transactions to bit /
sequence transactions as needed.

So now the app layer can extend TransactionGenerator, add it's own
custom messages and the existing TransactionConverter classes in use
in the sequences will deal with the conversion to the correct format as
required.