VERIFICATION: Reworked the "force_cid_change" arguments of the go_to_state tasks.
previously I had force_cid_change and cid arguments. If force_cid_change
was set, then the DUT was forced to re-enter the requested state with
the new CID.

force_cid_change defaulted to 0, and cid defaulted to $urandom_range(14)
(aka any valid CID). However it turns out that VCS (maybe based on the
verilog / SV LRM?) sets the default arguments at compile time. So the
$urandom_range is executed once during compilation, and not on every
call as I intended.

The new scheme moves the CidType enum from comms_tests_sequence.sv to
specific_target_sequence.sv, as well as a helper function. So now the
argument to the go_to_state tasks is a CidType, and we force re-entering
the requested state with a new Cid if cid_type is not CID_CURRENT.