VERIFICATION: Collect auto generated coverage and generate the report.
VCS can automatically collect coverage on: lines, conditions, FSMs,
toggles, branches and asserts. We collect all of that.

We can use this to show that our testbenches are doing a decent job of
testing the DUTs.

Note: We don't yet support functional coverage, as that requires
additional work to define all the relevant covergroups and coverpoints.