RTL: Fixed an issue with the tx_iface port to the crc_control module.
The crc_control module monitors the tx bus to decide when a new frame
starts and when to sample the data.

We were using the in_bit modport of the tx_interface, but that specifies
that req should be an output, however we read req and don't drive it.

The other option was the out_bit modport, but that specifies that data /
data_valid are outputs, and we read those too.

The solution is to create a new monitor_bit modport which specifies all
signals as inputs.