VERIFICATION: Update comments on AFE requirements and adjust default values.
The AFE must ensure that the clock stopping during a pause will not
cause the tag to miss more than 110 clock edges, as long as the format
of the PCD's pause is within the required range as defined by the
standard.

I updated the default values for clock_stops_after_ps and
clock_starts_after_ps to match values found from SPICE simulations of
Fabricio's AFE design. Although these values come from a design for
another fabrication process, they are the best we have for now.
Additionally the sequence_decode testbench shows that as long as these
pramaters are within the required range, my design will work.