RTL: Changing scheme for how pause_n latching and synchronisation works.
Previously I intended the pause_n signal to pass through a reset
synchroniser to latch it and synchronise it.

However I recently realised that a reset synchroniser doesn't
synchronise the asserting edge of the signal, since it expects that
signal to go to the asynchronous reset pin of other flip flops, rather
than the D pin.

Instead I'm using a single FFD to "latch" the pause signal. Pause_n is
connected to the async reset signal. Then the D pin is set to a constant
'1' and the clock goes to the clk pin. This means that when the pause
asserts the output (pause_n_latched) immediately asserts
(asynchronously). To then synchronise that signal a standard 2 stage
synchroniser is used.

When the pause_n input deasserts on the next rising edge of the clock
the '1' is clocked through, after two further rising clock edges, the
output of the synchroniser deasserts.

This way both the asserting and deasserting edges of the pause_n signal
are synchronous to our clock domain.