VERIFICATION: Fixed a bug with generating errors in the PCD pause driver
To test how my designs respond to errors we can force the drivers to
insert errors into the transaction. With the PCD's pause_n driver, we
have to do this by sending two pauses close together.

Previously I had the two pauses sent 48 ticks apart, which worked with
the old analogue_sim default values, but can be too large a gap if
pause_ticks is high and the analogue sim clock_starts/stops_after_ps
parameters are set such that only a few ticks clock edges are missed. In
that case the end of pauses are enough ticks apart that the sequence
doesn't detect it as an error.

I've calculated a new value for the gap between pauses which should
always be detected as an error no matter what the AFE or pause_ticks
values are (as long as they are in the required range).