VERIFICATION: We support the clock stopping for 59 ticks.
Previously I was using 58 as the maximum value, because the
dataValidOnlyOneTick assert failed in some cases for missing 59 ticks.
However I've realised that this assertion is not reflecting an error in
this case, as data_valid being asserted for two ticks in a row is a
possibility when receiving an XYZ series of sequences. The Y can time
out in one tick, and the Z be detected in the next.